+++ src/RobotInterface/include/BipedalLocomotion/RobotInterface/IRobotControl.h
@@ -147,6 +147,16 @@ public:
     virtual std::vector<std::string> getJointList() const = 0;
 
     /**
+     * Get the actuated joints limits.
+     * @param lowerLimits vector to be filled with the lower limits of the actuated joints.
+     * @param upperLimits vector to be filled with the upper limits of the actuated joints.
+     * @return True/False in case of success/failure.
+     */
+    virtual bool getJointLimits(Eigen::Ref<Eigen::VectorXd> lowerLimits,
+                                Eigen::Ref<Eigen::VectorXd> upperLimits) const
+        = 0;
+
+    /**
      * Check if the class is valid.
      * @note If it is valid you can directly control the robot
      * @return True if it is valid, false otherwise.